# Permit zero alloc reads